database/scripts/requirements.txt
# Python requirements for kanji processing script
psycopg2-binary==2.9.7
